always useLocationChange in usePath
